Adds descriptive comments to issue 3809 unit test.
The issue requires quite a specific set up in order to be reproduced,
and it could later be mistaken by a simpler scenario (where it would
work even without the fix). This information was added to the unit tests
to avoid simplifying it in the future, as Bridge.Html5 framework may at
some point be discontinued.